+++ src/middlewares/__init__.py
@@ -3,8 +3,13 @@ from telebot.handler_backends import BaseMiddleware
 
 from .register import RegisterMiddleware
 from .advert import AdvertMiddleware
+from .actives import ActiveMiddleware
 
 
-middlewares: list[Type[BaseMiddleware]] = [RegisterMiddleware, AdvertMiddleware]
+middlewares: list[Type[BaseMiddleware]] = [
+    RegisterMiddleware,
+    AdvertMiddleware,
+    ActiveMiddleware,
+]
 
 __all__ = ["middlewares"]